Pulled ResultSetMapper DSL support into a shared trait.
