+++ vanilla-core/src/main/groovy/com/stehno/vanilla/jdbc/mapper/transform/CompiledResultSetMapperBuilder.groovy
@@ -16,7 +16,7 @@
 package com.stehno.vanilla.jdbc.mapper.transform
 
 import com.stehno.vanilla.jdbc.mapper.FieldMapping
-import com.stehno.vanilla.jdbc.mapper.ResultSetMapperDsl
+import com.stehno.vanilla.jdbc.mapper.ResultSetMapperDslSupport
 import groovy.transform.CompileStatic
 import org.codehaus.groovy.ast.ClassNode
 
@@ -24,48 +24,16 @@ import org.codehaus.groovy.ast.ClassNode
  * Compile-time-based extension of the <code>ResultSetMapperBuilder</code> class. This class should not be used externally.
  */
 @CompileStatic
-class CompiledResultSetMapperBuilder implements ResultSetMapperDsl {
+class CompiledResultSetMapperBuilder implements ResultSetMapperDslSupport {
 
     final ClassNode mappedTypeNode
-
-    private final Collection<String> ignoredNames = []
-    private final Map<String, FieldMapping> mappings = [:]
 
     CompiledResultSetMapperBuilder(ClassNode mappedTypeNode) {
         this.mappedTypeNode = mappedTypeNode
     }
 
-    /**
-     * Maps the specified object property name and encapsulates it in a <code>FieldMapping</code>. The <code>FieldMapping</code> object is stored
-     * internally and a reference is returned.
-     *
-     * @param propertyName the object property being mapped
-     * @return a reference to the created FieldMapping object
-     */
-    FieldMapping map(String propertyName) {
-        FieldMapping mapping = new CompiledFieldMapping(propertyName)
-        mappings[propertyName] = mapping
-        mapping
-    }
-
-    /**
-     * Configures the specified object properties to be ignored by the mapping.
-     *
-     * @param propertyNames one or more object property names to be ignored
-     */
-    void ignore(String... propertyNames) {
-        this.ignoredNames.addAll(propertyNames)
-    }
-
-    /**
-     * Retrieves a collection containing all the configured mappings, with any ignored property mappings removed.
-     *
-     * @return a Collection of the field mappings
-     */
-    @SuppressWarnings('ConfusingMethodName')
-    Collection<FieldMapping> mappings() {
-        mappings.findAll { String fname, FieldMapping fm ->
-            !(fname in ignoredNames)
-        }.values().asImmutable()
+    @Override
+    FieldMapping createFieldMapping(String propertyName) {
+        new CompiledFieldMapping(propertyName)
     }
 }
